requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain_groq
langchain_huggingface
langchain_text_splitters
sentence_transformers
pypdf
faiss-cpu